An attempt to handle interface routing with c++ to match controllers directly with actors.
Turns out that compiling this stuff on mac is a
nightmare, and also this is probably the wrong implementation
of what I want. But I'll commit this for posterity.

Actual solution: a base modular character class that
handles routing via getComponentByClass.

The controller casts the pawn to that base class,
and the actors define how the functions are implemented.